test sample_extract

+++ src/fluid_dsp_float.c
@@ -10,7 +10,7 @@ extern  int16_t spi_read_int16(uint32_t address);
 
 #if defined(__arm__) && defined(SPI_READ_SAMPLE) 
 #define READ_SAMPLE(base, pos) \
-        ((base) > 0x08000000 ? base[pos] : SPI_READ_SAMPLE((base), (pos)))
+        ((int)(base) > 0x08000000 ? base[pos] : SPI_READ_SAMPLE((base), (pos)))
 #else
 #define READ_SAMPLE(base, pos) base[pos]
 #endif
